complete favorite category

+++ app/src/main/java/com/derlados/computer_conf/interfaces/BuildConstructorView.kt
@@ -2,7 +2,6 @@ package com.derlados.computer_conf.interfaces
 
 import android.graphics.Bitmap
 import com.derlados.computer_conf.consts.ComponentCategory
-import com.derlados.computer_conf.models.Build
 import com.derlados.computer_conf.models.BuildData
 import com.derlados.computer_conf.models.Component
 import com.derlados.computer_conf.presenters.BuildConstructorPresenter
@@ -14,7 +13,7 @@ interface BuildConstructorView {
     fun setStatus(status: BuildConstructorPresenter.StatusBuild, message: String? = null)
 
     fun addNewComponent(category: ComponentCategory, component: Component)
-    fun updatePrice(price: Float)
+    fun updatePrice(price: Int)
 
     fun showSaveDialog()
     fun exitView()
